Fixed some warning on Windows
git-svn-id: https://coolfluidsrv.vki.ac.be/svn/coolfluid3/trunk@197 4892e656-8fb9-448a-b5bc-b1919152596f

+++ src/Common/LogStream.cpp
@@ -336,8 +336,8 @@ void LogStream::removeStringForwarder(LogStringForwarder * forwarder)
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 //+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-int LogStream::getStringForwarderCount() const
+unsigned int LogStream::getStringForwarderCount() const
 {
-  return m_stringForwarders.size();
+  return (unsigned int) m_stringForwarders.size();
 }
